app/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
sympy>=1.12
numpy>=1.24.0
